Change for PIDs and TIDs in aggregators from UInt16Array to Float64Array (#1849)
Fixes: https://b.corp.google.com/issues/424815014

+++ ui/src/plugins/dev.perfetto.Sched/thread_state_selection_aggregator.ts
@@ -102,7 +102,7 @@ export class ThreadStateSelectionAggregator implements AreaSelectionAggregator {
       {
         title: 'PID',
         kind: 'NUMBER',
-        columnConstructor: Uint16Array,
+        columnConstructor: Float64Array,
         columnId: 'pid',
       },
       {
@@ -114,7 +114,7 @@ export class ThreadStateSelectionAggregator implements AreaSelectionAggregator {
       {
         title: 'TID',
         kind: 'NUMBER',
-        columnConstructor: Uint16Array,
+        columnConstructor: Float64Array,
         columnId: 'tid',
       },
       {
